FT-1256 Improve clarity of Fisher out. file.
Replace X and Y by more informative column names: use the name of the parent folder of the categorical variables as a column name, pasted together with a specification whether it was the independent or dependent variable, e.g. in the case of this example it should be “Independent variable - BRAF value” and “Dependent variable - KRAS value” (parent folder is BRAF, the categorical values are BRAF WT and BRAF mutated).